# dims-web/src/main/resources/api-docs.json
{
  "swagger": "2.0",
  "info": {
    "description": "更多Spring Boot相关文章请关注：http://blog.didispace.com/",
    "version": "1.0",
    "title": "Spring Boot中使用Swagger2构建RESTful APIs",
    "termsOfService": "http://blog.didispace.com/"
  },
  "host": "localhost:10001",
  "basePath": "/",
...
  "paths": {
    "/api/v1/filelocation": {
      "get": {
        "tags": [
          "file-config-rest-controller"
        ],
        "summary": "fileLocations",
        "operationId": "fileLocationsUsingGET",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "fileLocationId",
            "in": "query",
            "description": "fileLocationId",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "limit",
            "in": "query",
            "description": "limit",
            "required": false,
            "type": "integer",
            "default": 100,
            "format": "int32"
          },
          {
            "name": "name",
            "in": "query",
            "description": "name",
            "required": false,
            "type": "string"
          },
          {
            "name": "start",
            "in": "query",
            "description": "start",
            "required": false,
            "type": "integer",
            "default": 0,
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      },
      "post": {
        "tags": [
          "file-config-rest-controller"
        ],
        "summary": "saveFileLocation",
        "operationId": "saveFileLocationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "encoding",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "fileDelimiter",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "fileEncoding",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "fileLocationId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "host",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "name",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "password",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "path",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "port",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int32"
          },
          {
            "name": "schema",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "specialityName",
            "in": "query",
            "required": false,
            "type": "string"
          },
          {
            "name": "taskConfigId",
            "in": "query",
            "required": false,
            "type": "integer",
            "format": "int64"
          },
          {
            "name": "username",
            "in": "query",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/filelocation/{fileLocationId}": {
      "delete": {
        "tags": [
          "file-config-rest-controller"
        ],
        "summary": "deleteFileLocation",
        "operationId": "deleteFileLocationUsingDELETE",
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "fileLocationId",
            "in": "path",
            "description": "fileLocationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "204": {
            "description": "No Content"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/filelocation/{fileLocationId}/test": {
      "post": {
        "tags": [
          "file-config-rest-controller"
        ],
        "summary": "testFileLocation",
        "operationId": "testFileLocationUsingPOST",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "name": "fileLocationId",
            "in": "path",
            "description": "fileLocationId",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
    "/api/v1/taskconfig/configs": {
      "get": {
        "tags": [
          "task-config-rest-controller"
        ],
        "summary": "configs",
        "operationId": "configsUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/taskconfig/province": {
      "get": {
        "tags": [
          "task-config-rest-controller"
        ],
        "summary": "province",
        "operationId": "provinceUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
    "/api/v1/taskconfig/speciality": {
      "get": {
        "tags": [
          "task-config-rest-controller"
        ],
        "summary": "speciality",
        "operationId": "specialityUsingGET",
        "produces": [
          "*/*"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseDto"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        },
        "deprecated": false
      }
    },
...
  },
  "definitions": {
    "ResponseDto": {
      "type": "object",
      "properties": {
        "count": {
          "type": "integer",
          "format": "int32"
        },
        "data": {
          "type": "object"
        },
        "message": {
          "type": "string"
        },
        "status": {
          "type": "integer",
          "format": "int32"
        },
        "total": {
          "type": "integer",
          "format": "int64"
        }
      },
      "title": "ResponseDto"
    }
  }
}